checked PEP 8 of Flask architecture
